# packages/openapi-generator/tmp/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/posts/{postSlug}/comments": {
      "post": {
        "operationId": "CommentsController_createComment",
        "parameters": [
          {
            "name": "postSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  },
                  "parentId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  }
                },
                "required": ["content"],
                "title": "CreateCommentSchema",
                "description": "Schema for creating a comment"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentSchema"
                }
              }
            }
          }
        },
        "tags": ["Comments"]
      },
      "get": {
        "operationId": "CommentsController_getComments",
        "parameters": [
          {
            "name": "postSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Starting position of the query",
            "schema": {
              "default": 0,
              "description": "Starting position of the query",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "default": 20,
              "description": "Number of items to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Schema for sorting items",
            "schema": {
              "$ref": "#/components/schemas/CommentsController_getComments_sortArray"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "filter",
            "required": false,
            "in": "query",
            "description": "Filtering query string, in the format of \"property:rule[:value];property:rule[:value];...\"\nAvailable rules: eq, neq, gt, gte, lt, lte, like, nlike, in, nin, isnull, isnotnull \nAvailable properties: content",
            "schema": {
              "$ref": "#/components/schemas/CommentsController_getComments_filterArray"
            },
            "style": "form",
            "explode": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsSchema"
                }
              }
            }
          }
        },
        "tags": ["Comments"]
      }
    },
    "/api/posts/{postSlug}/comments/{commentId}": {
      "patch": {
        "operationId": "CommentsController_updateComment",
        "parameters": [
          {
            "name": "postSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "content": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1000
                  }
                },
                "required": ["content"],
                "title": "UpdateCommentSchema",
                "description": "Schema for updating a comment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentSchema"
                }
              }
            }
          }
        },
        "tags": ["Comments"]
      },
      "delete": {
        "operationId": "CommentsController_deleteComment",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "postSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Comments"]
      }
    },
    "/api/posts/{postSlug}/comments/count": {
      "get": {
        "operationId": "CommentsController_getCommentCount",
        "parameters": [
          {
            "name": "postSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": ["Comments"]
      }
    },
    "/api/posts/{postSlug}/comments/{commentId}/replies": {
      "get": {
        "operationId": "CommentsController_getCommentReplies",
        "parameters": [
          {
            "name": "commentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Starting position of the query",
            "schema": {
              "default": 0,
              "description": "Starting position of the query",
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Number of items to return",
            "schema": {
              "default": 20,
              "description": "Number of items to return",
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "Schema for sorting items",
            "schema": {
              "$ref": "#/components/schemas/CommentsController_getCommentReplies_sortArray"
            },
            "style": "form",
            "explode": false
          },
          {
            "name": "postSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentsSchema"
                }
              }
            }
          }
        },
        "tags": ["Comments"]
      }
    },
...
  },
  "info": {
    "title": "Lonestone API",
    "description": "The Lonestone API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CommentSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "content": {
            "type": "string"
          },
          "authorName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "user": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "name": {
                    "type": "string"
                  }
                },
                "required": ["id", "name"],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "parentId": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              {
                "type": "null"
              }
            ]
          },
          "replyIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          },
          "replyCount": {
            "type": "number"
          }
        },
        "required": ["id", "content", "authorName", "createdAt", "user", "parentId"],
        "additionalProperties": false,
        "title": "CommentSchema",
        "description": "Schema for a comment"
      },
...
      "CommentsSchema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CommentSchema"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "offset": {
                "type": "number"
              },
              "pageSize": {
                "type": "number"
              },
              "itemCount": {
                "type": "number"
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": ["offset", "pageSize", "itemCount", "hasMore"],
            "additionalProperties": false
          }
        },
        "required": ["data", "meta"],
        "additionalProperties": false,
        "title": "CommentsSchema",
        "description": "Schema for a paginated list of comments"
      },
...
      "CommentsController_getComments_sortItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["createdAt", "authorName"]
          },
          "direction": {
            "type": "string",
            "enum": ["asc", "desc"]
          }
        },
        "required": ["property", "direction"],
        "additionalProperties": false
      },
      "CommentsController_getComments_sortArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/CommentsController_getComments_sortItem"
        }
      },
      "CommentsController_getComments_filterItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["content"]
          },
          "rule": {
            "type": "string",
            "enum": [
              "eq",
              "neq",
              "gt",
              "gte",
              "lt",
              "lte",
              "like",
              "nlike",
              "in",
              "nin",
              "isnull",
              "isnotnull"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": ["property", "rule"],
        "additionalProperties": false
      },
      "CommentsController_getComments_filterArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/CommentsController_getComments_filterItem"
        }
      },
      "CommentsController_getCommentReplies_sortItem": {
        "type": "object",
        "properties": {
          "property": {
            "type": "string",
            "enum": ["createdAt", "authorName"]
          },
          "direction": {
            "type": "string",
            "enum": ["asc", "desc"]
          }
        },
        "required": ["property", "direction"],
        "additionalProperties": false
      },
      "CommentsController_getCommentReplies_sortArray": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/CommentsController_getCommentReplies_sortItem"
        }
      },
...
    }
  }
}